better separate crypto from protocol logic
